Create BirthdayParty
For her daughter's birthday, Ludmila has decided to organize a party to which she invites all her classmates. For this purpose, she decided to rent an entertainment room for children, whose rent you will receive as an entrance from the console.
Write a program to help Ludmila calculate what budget she will need, with the following information about the extra things needed for the celebration:
• Cake - its price is 20% of the rent of the hall
• Drinks - their price is 45% less than that of the cake
• Animator - its price is 1/3 of the price for renting the hall

1 line is read from the console:
• Rent for the hall - real number
To print on the console what budget will be needed to organize the celebration.