Create Vacation Book List
There are a number of books on Joro's list of required literature for the summer vacation, but Joro prefers to play with friends outside. Your task is to help Joro calculate how many hours a day he has to spend to read the necessary literature, but also to spend maximum time outside.