Create Charity Campaign
The confectionery is running a charity fundraising campaign, which can involve confectioners from all over the country. Initially, we read from the console the number of days in which the campaign runs and the number of confectioners who will participate. Then in separate rows we get the amount of cakes, waffles and pancakes, which will be prepared by a confectioner in one day. The following price list must be kept in mind:
• Cake - BGN 45.
• Waffles - BGN 5.80.
• Pancake - BGN 3.20.
1/8 of the final amount will be used to cover the cost of the products during the campaign. Write a program that calculates the amount collected at the end of the campaign.