Create Fruit Market
Maria decides to go on a diet and goes to a nearby market to buy strawberries, bananas, oranges and raspberries. The price of strawberries in BGN / kg is entered on the console. and the amount of bananas, oranges, raspberries and strawberries he needs to buy. Write a program that calculates how much money she needs to pay the bill, knowing that:
• the price of raspberries is half lower than that of strawberries;
• the price of oranges is 40% lower than the price of raspberries;
• the price of bananas is 80% lower than the price of raspberries.